Move agent transcript parsing into agent-session
