+++ agent-session/Cargo.toml
@@ -15,5 +15,7 @@ categories = ["development-tools", "parser-implementations"]
 [dependencies]
 chrono = "0.4"
 dirs = "6"
+hex = "0.4"
 serde = { version = "1.0.219", features = ["derive"] }
 serde_json = "1.0.140"
+sha2 = "0.10"
