+++ agentpprof/Cargo.toml
@@ -18,16 +18,13 @@ agent-session = { version = "0.4.0", path = "../agent-session" }
 chrono = { version = "0.4", features = ["serde"] }
 clap = { version = "4.5", features = ["derive"] }
 dirs = "6"
-hex = "0.4"
 serde = { version = "1", features = ["derive"] }
 serde_json = "1"
 sha2 = "0.10"
-shell-words = "1"
 flate2 = "1"
 prost = "0.13"
 regex = "1"
 ureq = { version = "2", features = ["json"] }
-walkdir = "2"
 
 [dev-dependencies]
 tempfile = "3"
